+++ enjoy/src/renderer/components/chats/chat-user-message.tsx
@@ -1,0 +1,204 @@
+import {
+  Avatar,
+  AvatarFallback,
+  AvatarImage,
+  Button,
+  Collapsible,
+  CollapsibleContent,
+  CollapsibleTrigger,
+  toast,
+} from "@renderer/components/ui";
+import {
+  MarkdownWrapper,
+  PronunciationAssessmentScoreDetail,
+  WavesurferPlayer,
+} from "@renderer/components";
+import { formatDateTime } from "@renderer/lib/utils";
+import { t } from "i18next";
+import {
+  ChevronDownIcon,
+  ChevronUpIcon,
+  DownloadIcon,
+  GaugeCircleIcon,
+  LoaderIcon,
+  MicIcon,
+  SendIcon,
+  SparkleIcon,
+} from "lucide-react";
+import { useContext, useEffect, useRef, useState } from "react";
+import {
+  AppSettingsProviderContext,
+  ChatSessionProviderContext,
+} from "@renderer/context";
+import { useAiCommand } from "@renderer/hooks";
+import { md5 } from "js-md5";
+
+export const ChatUserMessage = (props: { chatMessage: ChatMessageType }) => {
+  const { chatMessage } = props;
+  const { EnjoyApp } = useContext(AppSettingsProviderContext);
+  const {
+    chatMessages,
+    askAgent,
+    startRecording,
+    isRecording,
+    isPaused,
+    setAssessing,
+  } = useContext(ChatSessionProviderContext);
+  const { recording } = chatMessage;
+  const ref = useRef<HTMLDivElement>(null);
+  const [suggestion, setSuggestion] = useState<string>();
+  const [suggesting, setSuggesting] = useState<boolean>(false);
+  const [suggestionVisible, setSuggestionVisible] = useState<boolean>(true);
+  const { refine } = useAiCommand();
+
+  const handleSuggest = async (params?: { reload?: boolean }) => {
+    if (suggesting) return;
+    if (!chatMessage.content) return;
+
+    const { reload = false } = params || {};
+    const cacheKey = `chat-message-suggestion-${md5(chatMessage.id)}`;
+    try {
+      const cached = await EnjoyApp.cacheObjects.get(cacheKey);
+
+      if (cached && !reload && !suggestion) {
+        setSuggestion(cached);
+      } else {
+        setSuggesting(true);
+
+        const context = `I'm chatting in a chatroom. The previous messages are as follows:\n\n${buildChatHistory()}`;
+        const result = await refine(chatMessage.content, context);
+        EnjoyApp.cacheObjects.set(cacheKey, result);
+        setSuggestion(result);
+        setSuggesting(false);
+      }
+    } catch (err) {
+      toast.error(err.message);
+      setSuggesting(false);
+    }
+  };
+
+  const buildChatHistory = () => {
+    const messages = chatMessages.filter(
+      (m) => new Date(m.createdAt) < new Date(chatMessage.createdAt)
+    );
+    return messages
+      .map(
+        (message) =>
+          `${(message.member.user || message.member.agent).name}: ${
+            message.content
+          }`
+      )
+      .join("\n");
+  };
+
+  useEffect(() => {
+    if (ref.current) {
+      ref.current.scrollIntoView({ behavior: "smooth" });
+    }
+  }, [ref]);
+
+  return (
+    <div ref={ref} className="mb-6">
+      <div className="flex items-center space-x-2 justify-end mb-2">
+        <div className="text-sm text-muted-foreground">
+          {chatMessage.member.user.name}
+        </div>
+        <Avatar className="w-8 h-8 bg-background avatar">
+          <AvatarImage src={chatMessage.member.user.avatarUrl}></AvatarImage>
+          <AvatarFallback className="bg-background">
+            {chatMessage.member.user.name}
+          </AvatarFallback>
+        </Avatar>
+      </div>
+      <div className="flex justify-end">
+        <div className="flex flex-col gap-2 px-4 py-2 mb-2 bg-sky-500/30 border-sky-500 rounded-lg shadow-sm w-full max-w-3xl">
+          {recording && (
+            <WavesurferPlayer id={recording.id} src={recording.src} />
+          )}
+          {recording?.pronunciationAssessment && (
+            <div className="flex justify-end">
+              <PronunciationAssessmentScoreDetail
+                assessment={recording.pronunciationAssessment}
+              />
+            </div>
+          )}
+          <MarkdownWrapper className="select-text prose dark:prose-invert">
+            {chatMessage.content}
+          </MarkdownWrapper>
+          {suggestion && (
+            <Collapsible
+              open={suggestionVisible}
+              onOpenChange={(value) => setSuggestionVisible(value)}
+            >
+              <CollapsibleContent>
+                <div className="p-4 font-serif bg-background rounded">
+                  <MarkdownWrapper className="select-text prose dark:prose-invert">
+                    {suggestion}
+                  </MarkdownWrapper>
+                </div>
+              </CollapsibleContent>
+              <div className="my-2 flex justify-center">
+                <CollapsibleTrigger asChild>
+                  <Button
+                    onClick={() => setSuggestionVisible(!suggestionVisible)}
+                    variant="ghost"
+                    size="icon"
+                    className="w-6 h-6"
+                  >
+                    {suggestionVisible ? (
+                      <ChevronUpIcon className="w-4 h-4" />
+                    ) : (
+                      <ChevronDownIcon className="w-4 h-4" />
+                    )}
+                  </Button>
+                </CollapsibleTrigger>
+              </div>
+            </Collapsible>
+          )}
+          <div className="flex items-center space-x-4">
+            {suggesting ? (
+              <LoaderIcon className="w-4 h-4 animate-spin" />
+            ) : (
+              <SparkleIcon
+                data-tooltip-id="global-tooltip"
+                data-tooltip-content={t("suggestion")}
+                className="w-4 h-4 cursor-pointer"
+                onClick={() => handleSuggest()}
+              />
+            )}
+            {chatMessage.state === "pending" ? (
+              <>
+                {isPaused || isRecording ? (
+                  <LoaderIcon className="w-4 h-4 animate-spin" />
+                ) : (
+                  <MicIcon
+                    data-tooltip-id="global-tooltip"
+                    data-tooltip-content={t("reRecord")}
+                    className="w-4 h-4 cursor-pointer"
+                    onClick={startRecording}
+                  />
+                )}
+                <SendIcon
+                  data-tooltip-id="global-tooltip"
+                  data-tooltip-content={t("confirm")}
+                  className="w-4 h-4 cursor-pointer"
+                  onClick={() => askAgent()}
+                />
+              </>
+            ) : (
+              <GaugeCircleIcon
+                data-tooltip-id="global-tooltip"
+                data-tooltip-content={t("pronunciationAssessment")}
+                onClick={() => setAssessing(recording)}
+                className="w-4 h-4 cursor-pointer"
+              />
+            )}
+          </div>
+        </div>
+      </div>
+      <div className="flex justify-end text-xs text-muted-foreground timestamp">
+        {formatDateTime(chatMessage.createdAt)}
+      </div>
+    </div>
+  );
+};
